C++ simulated detector

+++ core-buffer/include/jungfrau.hpp
@@ -17,7 +17,7 @@ struct jungfrau_packet {
     uint32_t exptime;
     uint32_t packetnum;
 
-    double bunchid;
+    uint64_t bunchid;
     uint64_t timestamp;
 
     uint16_t moduleID;
